@@ -10,10 +10,10 @@
     "@testing-library/user-event": "^7.1.2",
     "bootstrap": "^5.2.3",
     "jwt-decode": "^2.2.0",
-    "react": "^16.13.1",
+    "react": "^17.0.1",
     "react-bootstrap": "^2.7.0",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2",
+    "react-dom": "^17.0.1",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.1"
   },
   "scripts": {
